ln-resource-mgr/refactor: move revenue tracking into incoming channel
